fix: Merge `Hide autoplay button`
