+++ src/main/kotlin/app/revanced/patches/youtube/layout/buttons/autoplay/HideAutoplayButtonPatch.kt
@@ -1,84 +1,14 @@
 package app.revanced.patches.youtube.layout.buttons.autoplay
 
 import app.revanced.patcher.data.BytecodeContext
-import app.revanced.patcher.extensions.InstructionExtensions.addInstructionsWithLabels
-import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
 import app.revanced.patcher.patch.BytecodePatch
-import app.revanced.patcher.patch.annotation.CompatiblePackage
-import app.revanced.patcher.patch.annotation.Patch
-import app.revanced.patcher.util.smali.ExternalLabel
-import app.revanced.patches.all.misc.resources.AddResourcesPatch
-import app.revanced.patches.shared.misc.mapping.ResourceMappingPatch
-import app.revanced.patches.shared.misc.settings.preference.SwitchPreference
-import app.revanced.patches.youtube.misc.integrations.IntegrationsPatch
-import app.revanced.patches.youtube.misc.settings.SettingsPatch
-import app.revanced.patches.youtube.shared.fingerprints.LayoutConstructorFingerprint
-import app.revanced.util.getReference
-import app.revanced.util.indexOfFirstInstructionOrThrow
-import app.revanced.util.indexOfIdResourceOrThrow
-import app.revanced.util.resultOrThrow
-import com.android.tools.smali.dexlib2.Opcode
-import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
-import com.android.tools.smali.dexlib2.iface.reference.MethodReference
+import app.revanced.patches.youtube.layout.buttons.player.HidePlayerOverlayButtonsPatch
 
-@Patch(
-    name = "Hide autoplay button",
-    description = "Adds an option to hide the autoplay button in the video player.",
-    dependencies = [
-        IntegrationsPatch::class,
-        SettingsPatch::class,
-        ResourceMappingPatch::class,
-        AddResourcesPatch::class,
-    ],
-    compatiblePackages = [
-        CompatiblePackage(
-            "com.google.android.youtube",
-            [
-                "18.38.44",
-                "18.49.37",
-                "19.16.39",
-                "19.25.37",
-                "19.34.42",
-            ],
-        ),
-    ],
-)
 @Suppress("unused")
+@Deprecated("This patch has been merged into HidePlayerOverlayButtonsPatch.")
 object HideAutoplayButtonPatch : BytecodePatch(
-    setOf(LayoutConstructorFingerprint),
+    dependencies = setOf(HidePlayerOverlayButtonsPatch::class),
 ) {
-
-    private const val INTEGRATIONS_CLASS_DESCRIPTOR =
-        "Lapp/revanced/integrations/youtube/patches/HideAutoplayButtonPatch;"
-
     override fun execute(context: BytecodeContext) {
-        AddResourcesPatch(this::class)
-
-        SettingsPatch.PreferenceScreen.PLAYER.addPreferences(
-            SwitchPreference("revanced_hide_autoplay_button"),
-        )
-
-        LayoutConstructorFingerprint.resultOrThrow().mutableMethod.apply {
-            val constIndex = indexOfIdResourceOrThrow("autonav_toggle")
-            val constRegister = getInstruction<OneRegisterInstruction>(constIndex).registerA
-
-            // Add a conditional branch around the code that inflates and adds the auto repeat button.
-            val gotoIndex = indexOfFirstInstructionOrThrow(constIndex) {
-                val parameterTypes = getReference<MethodReference>()?.parameterTypes
-                opcode == Opcode.INVOKE_VIRTUAL &&
-                    parameterTypes?.size == 2 &&
-                    parameterTypes.first() == "Landroid/view/ViewStub;"
-            } + 1
-
-            addInstructionsWithLabels(
-                constIndex,
-                """
-                    invoke-static {}, $INTEGRATIONS_CLASS_DESCRIPTOR->hideAutoPlayButton()Z
-                    move-result v$constRegister
-                    if-nez v$constRegister, :hidden
-                """,
-                ExternalLabel("hidden", getInstruction(gotoIndex)),
-            )
-        }
     }
 }
