+++ src/main/kotlin/app/revanced/patches/youtube/layout/buttons/player/HidePlayerOverlayButtonsPatch.kt
@@ -3,23 +3,29 @@ package app.revanced.patches.youtube.layout.buttons.player
 import app.revanced.patcher.data.BytecodeContext
 import app.revanced.patcher.extensions.InstructionExtensions.addInstruction
 import app.revanced.patcher.extensions.InstructionExtensions.addInstructions
+import app.revanced.patcher.extensions.InstructionExtensions.addInstructionsWithLabels
 import app.revanced.patcher.extensions.InstructionExtensions.getInstruction
 import app.revanced.patcher.patch.BytecodePatch
 import app.revanced.patcher.patch.PatchException
 import app.revanced.patcher.patch.annotation.CompatiblePackage
 import app.revanced.patcher.patch.annotation.Patch
+import app.revanced.patcher.util.smali.ExternalLabel
 import app.revanced.patches.all.misc.resources.AddResourcesPatch
 import app.revanced.patches.shared.misc.settings.preference.SwitchPreference
+import app.revanced.patches.youtube.layout.buttons.autoplay.HideAutoplayButtonPatch
 import app.revanced.patches.youtube.layout.buttons.player.fingerprints.PlayerControlsPreviousNextOverlayTouchFingerprint
 import app.revanced.patches.youtube.misc.integrations.IntegrationsPatch
 import app.revanced.patches.youtube.misc.settings.SettingsPatch
+import app.revanced.patches.youtube.shared.fingerprints.LayoutConstructorFingerprint
 import app.revanced.patches.youtube.shared.fingerprints.SubtitleButtonControllerFingerprint
 import app.revanced.util.getReference
 import app.revanced.util.indexOfFirstInstructionOrThrow
 import app.revanced.util.indexOfFirstWideLiteralInstructionValueOrThrow
+import app.revanced.util.indexOfIdResourceOrThrow
 import app.revanced.util.resultOrThrow
 import com.android.tools.smali.dexlib2.Opcode
 import com.android.tools.smali.dexlib2.iface.instruction.FiveRegisterInstruction
+import com.android.tools.smali.dexlib2.iface.instruction.OneRegisterInstruction
 import com.android.tools.smali.dexlib2.iface.reference.MethodReference
 
 @Patch(
@@ -48,7 +54,8 @@ import com.android.tools.smali.dexlib2.iface.reference.MethodReference
 object HidePlayerOverlayButtonsPatch : BytecodePatch(
     setOf(
         PlayerControlsPreviousNextOverlayTouchFingerprint,
-        SubtitleButtonControllerFingerprint
+        SubtitleButtonControllerFingerprint,
+        LayoutConstructorFingerprint
     )
 ) {
 
@@ -113,6 +120,33 @@ object HidePlayerOverlayButtonsPatch : BytecodePatch(
             addInstruction(
                 insertIndex,
                 "invoke-static {v0}, $INTEGRATIONS_CLASS_DESCRIPTOR->hideCaptionsButton(Landroid/widget/ImageView;)V"
+            )
+        }
+
+        // endregion
+
+        // region hide auto play button
+
+        LayoutConstructorFingerprint.resultOrThrow().mutableMethod.apply {
+            val constIndex = indexOfIdResourceOrThrow("autonav_toggle")
+            val constRegister = getInstruction<OneRegisterInstruction>(constIndex).registerA
+
+            // Add a conditional branch around the code that inflates and adds the auto repeat button.
+            val gotoIndex = indexOfFirstInstructionOrThrow(constIndex) {
+                val parameterTypes = getReference<MethodReference>()?.parameterTypes
+                opcode == Opcode.INVOKE_VIRTUAL &&
+                        parameterTypes?.size == 2 &&
+                        parameterTypes.first() == "Landroid/view/ViewStub;"
+            } + 1
+
+            addInstructionsWithLabels(
+                constIndex,
+                """
+                    invoke-static {}, $INTEGRATIONS_CLASS_DESCRIPTOR->hideAutoPlayButton()Z
+                    move-result v$constRegister
+                    if-nez v$constRegister, :hidden
+                """,
+                ExternalLabel("hidden", getInstruction(gotoIndex)),
             )
         }
