+++ src/main/resources/addresources/values/strings.xml
@@ -430,11 +430,6 @@ This is because Crowdin requires temporarily flattening this file and removing t
             <string name="revanced_hide_playlist_button_summary_on">Save to playlist button is hidden</string>
             <string name="revanced_hide_playlist_button_summary_off">Save to playlist button is shown</string>
         </patch>
-        <patch id="layout.buttons.autoplay.HideAutoplayButtonPatch">
-            <string name="revanced_hide_autoplay_button_title">Hide autoplay button</string>
-            <string name="revanced_hide_autoplay_button_summary_on">Autoplay button is hidden</string>
-            <string name="revanced_hide_autoplay_button_summary_off">Autoplay button is shown</string>
-        </patch>
         <patch id="layout.buttons.navigation.NavigationButtonsPatch">
             <string name="revanced_navigation_buttons_screen_title">Navigation buttons</string>
             <string name="revanced_navigation_buttons_screen_summary">Hide or change buttons in the navigation bar</string>
@@ -518,6 +513,9 @@ This is because Crowdin requires temporarily flattening this file and removing t
             <string name="revanced_hide_captions_button_title">Hide captions button</string>
             <string name="revanced_hide_captions_button_summary_on">Captions button is hidden</string>
             <string name="revanced_hide_captions_button_summary_off">Captions button is shown</string>
+            <string name="revanced_hide_autoplay_button_title">Hide autoplay button</string>
+            <string name="revanced_hide_autoplay_button_summary_on">Autoplay button is hidden</string>
+            <string name="revanced_hide_autoplay_button_summary_off">Autoplay button is shown</string>
         </patch>
         <patch id="layout.hide.albumcards.AlbumCardsResourcePatch">
             <string name="revanced_hide_album_cards_title">Hide album cards</string>
